Week 12 DOM selectors

+++ 025-DOM-selectors/main.js
@@ -21,9 +21,9 @@ data as an object into global array named MY_DATA */
 
 FORM.addEventListener('submit', (e) => {
     e.preventDefault();
-    const miles = parseInt(e.target.miles.value);
-    const gallons = parseInt(e.target.gallons.value);
-    const price = parseInt(e.target.price.value);
+    const miles = parseFloat(e.target.miles.value);
+    const gallons = parseFloat(e.target.gallons.value);
+    const price = parseFloat(e.target.price.value);
     const isValid = isFormValid(miles, gallons, price);
     if (isValid) {
         ERR.textContent = '';
